backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-multipart>=0.0.6
langgraph>=0.0.23
langchain>=0.1.0
langchain-core>=0.1.16
langchain-ollama>=0.1.0
paramiko>=3.4.0
requests>=2.31.0
pyyaml>=6.0.1
python-jose>=3.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
mcp>=1.0.0

